Crate pmemobj_sys

FFI bindings to libpmemobj

The libpmemobj library provides a transactional object store, providing memory allocation, transactions, and general facilities for persistent memory programming.

Developers new to persistent memory probably want to start with this library.

This is not an official port of the NVM Library.

The official libpmemobj documentation can be found at: http://pmem.io/nvml/libpmemobj/

Enums

PMEMobjpool

Functions

pmemobj_close

pmemobj_create

pmemobj_drain

pmemobj_errormsg

pmemobj_flush

pmemobj_memcpy_persist

pmemobj_memset_persist

pmemobj_open

pmemobj_persist